requirements.txt
boto3>=1.34.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
prometheus-client>=0.19.0
pyyaml>=6.0
python-multipart>=0.0.6
jinja2>=3.1.0
botocore>=1.34.0